Animation of the first scene has been fixed
Now it is ensured that the girl will move when the dad looks towards her, and as a safety measure, a simple event has been placed in the hole, to ensure that when the dad comes near the girl, the girl will start moving.

The repeat animation has been fixed, thus, they will repeat and not just play once.