Fixed Scene 2 door sound conflict
When the player triggers the door, the sounds plays depending on whether the key entered the area before the player body. If the key entered first, everything will work fine, if the player entered first, a sequence of sounds will play to give a hint for the player that they need to look for a key. The problem is that the sequence continue playing even when the key enter the door trigger area and the end of scene event is triggered. This has now been fixed. When the key enter the trigger, it interrupts the sequence and stop it.