Bug in scene 1 fixed
There was a bug that caused the game to crash when the girl went from waiting at the last door in the long corridor to the second stage where she is inactive. The cause of the problem is the AnimateListener on the event at Destination 5. This script waited for five seconds before calling for the next sound to be played, however, in order to get the sound to play, a coroutine using a component script attached to the girl character was used. Because the girl character was inactive, the coroutine could not be started. To fix the problem an if statement to check whether the girl is active and enabled or not has been added.